Move Buildtest global fixture into plugin
My original plan to stick this in our monorepo's parent conftest.py has
met quite a bit of friction because we do so many creative things with
pyproject.toml and pytest.ini throughout the monorepo. In short, the
existence of either configuration lower on the path prevents the
discovery of the parent pyproject.toml or confest.py in the root
directory.

There are a couple of different ways we can address this:

- We can remove child directory configs
- We can turn our Buildkite Test Engine checking into a plugin and
  register that plugin with all tests

Neither option is perfect - both rely on convention to make sure this is
configured on new test suites. We're heavy on convention today - lots of
our tox.ini files are cargo-culted around. I'd like to get this broadly
working for the existing test suite and then will address how to ensure
it stays working as test suites get added or changed.
