+++ python_modules/dagster/tox.ini
@@ -15,6 +15,7 @@ passenv =
     BUILDKITE*
     DAGSTER_DOCKER_*
     GRPC_SERVER_HOST
+    PYTEST_PLUGINS
 install_command =
   ; pyright/mypy invocations type_signature_tests in need commpat mode installs to find types correctly
   type_signature_tests: uv pip install {opts} {packages} --config-settings editable_mode=compat
@@ -27,9 +28,10 @@ deps =
   components_tests: -e ../libraries/dagster-dg
   components_tests: -e .[test-components]
   -e ../dagster-test
-  -e .[mypy,test,pyright]
+  -e .[mypy,test,pyright,ci]
   -e ../dagster-pipes
   -e ../libraries/dagster-shared
+  -e ../../.buildkite/dagster-buildkite
 
   # These are required for dagster dev tests, but we leave them out of `dagster[test]` so that
   # all of our other libs that depend on dagster[test] don't have to install them.
